/api/v1/users/me/my-page-info rest api

+++ src/main/java/com/gaji/app/member/domain/Member.java
@@ -39,5 +39,5 @@ public class Member {
     private String profilePic;
 
     @Column(name = "MANNERTEMP", nullable = false)
-    private Double mannerTemp;
+    private Integer mannerTemp;
 }
